Big set of changes.
Move TraceLine to world.c in nq, and added to pmovetst.c in qw.
Improved traceline to give the normal as well.

GL_LoadTexture now takes an additional argument, bpp.
For all existing cases it is 8, it may also be 32.

Tweaked the rocket trails, longer cones, fewer of them.

Reduced the number of particles from an explosion from 1024 to 512.

Grabbed the explosion effect from DP for qw, but I have it disabled.
(It looks nice, but is much slower then the particle explosions now.)