Changes for NQ.  QW may follow, but because of the differences in protocols, it won't be easily ported.  Added Cvars sv_jumpstep (0) and sv_stepheight (18) for BloodBath.  Set sv_jumpstep to 1 if you want to be able to climb stairs while jumping.  Quake2 and beyond have this, as does any engine which has implemented the hacky QW fly movement (we haven't, so it doesn't apply to us..)
Modified the NQ physics a bit to make objects that would normally have
fallen outside the map ... not do so.  It's still possible to _get_
outside the map through engine or QC bugs, but when that happens to a
player, he's given temporary MOVETYPE_NOCLIP.  This will allow him to at
least get back into the map, though it's possible for the unscrupulous to
use it as a cheat.  Of course, they have to find such a bug and exploit it
first.  =)

NQ now has SV_FreeMove, a function which mimics the control of water
movement without the friction, accelleration, etc.  If you look down and
press forward, you'll actually move down.  This is now used for NOCLIP and
FLY movements.  Also, while flying you will automatically climb stairs as
if sv_jumpstep were enabled.