Progs tweaking.  Much diff reduction has been started and not finished. This means lots of whitespace changes that will probably need changing again just because I don't have time to work more on this right now. Sorry 'bout that.  We now support builtin 99, checkextension, in both NQ and QW.  Here's the currently supported extension list: DP_QC_CHANGEPITCH, DP_QC_ETOS, DP_QC_FINDFLOAT, DP_QC_MINMAXBOUND, DP_QC_RANDOMVEC, DP_QC_SINCOSSQRTPOW, DP_QC_TRACEBOX, and DP_REGISTERCVAR.  None of these conflict with QW.
Note to anyone working on progs, QW and NQ progs are (or easily can be)
essentially identical with the exception of QW's string handling and the
builtin commands.  While not trivial, QW's string handling should be
backported to NQ for 64 bit sanity anyway.  Merging the commands
themselves require merging of things like edicts.  I don't think we're
ready for that quite yet.