WARNING: This commit breaks things you'd not expect it to break!
The nq, qw, and include directories have moved in CVS from $(top_srcdir)
to $(top_srcdir)/src as a prelude to actually allowing you to sanely use
your source directory as a game directory.  You'll need to ./bootstrap by
hand, and if you have an outstanding patch, you'll want to re-apply it by
hand.  There are lots of ways to do it, but most of them are at least a
minor annoyance.

This is just the structure, the real work in the code I want to wait until
after Zeph adjusts to the structure changes since he's got some major
outstanding code in his working dir.