Before I do more, I should commit this much.
Kill light.h, now part of gl_light.h.

We now REQUIRE SDL 1.2.5.

Interduce sdlimage.[ch], if we can find SDL_image, use it to load
things.

(Still a hack in a few ways, like the init, but we get /lots/ of image
formats from this.)

Interduce VectorMI, like VectorMA, except that it does a += on the
destination.

Do some tweaking on the alias model renderer:
 - Kill the shadelight variable that was always 1.
 - Move lightcolor to gl_alias.c.
 - Make the blended alias model function use VectorCopy, and VectorMI.

Change R_LightPoint to take a vec3_t argument, set to the light color
of the point.

Probably a few other things.