Change the search order for SDL loaded images.
Move the host_basepal and host_colormap loading to
src/renderer/palette.c.

If we can't find the colormap, print a warning and assaume 32 colors.

Remove gfx/palette.lmp from the embedded_data.

Change the font a bit, it is now 256x256, and shipped as a tga. (because
if we can't find the gamedata, we probably can't find SDL_Image.)

Also ship the .xcf for others.

VID_Init no longer takes an argument. (It was not used.)