The 'well THAT needs fixing!' commit.
Plug two rather sizable file descriptor leaks.

Change the qw client 'going too fast' checking to waste /far/ less CPU
time, this will at best be seen as being able to do more on a client
that can easily manage the requested FPS.
(Quite possibly making the framerate more stable!)

Add show_fps 2, which adds two numbers to the FPS counter, the first is
the number of times this frame we have spent looping because we are
running too fast.  The second is the number of times we have used
SDL_Delay to avoid wasting CPU time while we spin, due to going too
fast.

Ideally, on a really fast system, the FPS counter should show the
requested FPS, and the second number should be 1, with the first number
being very low.